Refactor ProxyUnchain into Util\ProxyUnchain
